v1.0.21: fix 9 pipeline bugs, remove leaked credentials, build fixes
Security:
- Remove build_installers.bat, sign_installers.bat, sign_and_pack.bat
  (contained plaintext code-signing passwords in git history).
  Add *.pfx, sign_*.bat, build_installer*.bat to .gitignore.

Pipeline bug fixes:
- Fix off-by-one in DecodedText: left(i1-1) → left(i1) lost last char
  before \r in decoded messages.
- Fix DxCluster bandFromFreq(0) mapping to "160M" instead of "UNK";
  add guard freqKhz < 1800 and early return on freq <= 0 in parseSpotLine.
- Fix signed-int overflow on m_asyncAudioPos (widgets path): int → uint64_t
  to prevent UB after ~50 hours of continuous operation.
- Fix UDP port truncation: bounds-check UDPServerPort/UDPListenPort values
  > 65535 with fallback to defaults.
- Remove hardcoded debug fopen("C:\Users\IU8LMC\...") from FtxFt8Stage4.cpp
  and residual comment from DecodiumCatManager.cpp.

CI/CD:
- Add concurrency groups to all 9 GitHub workflows to prevent parallel
  build races. Release workflows use cancel-in-progress: false.

Build fixes:
- Fix map65/libm65/ptt.c: CreateFile(TEXT(s)) → CreateFileA(s) for GCC 15.
- Fix main.cpp: remove deprecated QSysInfo::WV_VISTA (removed in Qt 6.6).

Version:
- Bump fork_release_version.txt to 1.0.21.
- Update installer/decodium_setup.iss: AppVersion 1.0.10 → 1.0.21,
  fix AppPublisherURL from elisir80 to iu8lmc.

+++ widgets/mainwindow.cpp
@@ -3499,8 +3499,8 @@ MainWindow::MainWindow(QDir const& temp_directory, bool multiple,
 
     decodium::ft2::AsyncDecodeRequest request;
     request.audio.resize (45000);
-    int pos = m_asyncAudioPos;
-    int start = (pos - 45000 + 90000) % 90000;
+    uint64_t pos = m_asyncAudioPos;
+    int start = static_cast<int>((pos - 45000 + 90000) % 90000);
     for (int i = 0; i < 45000; i++) {
       request.audio[i] = m_asyncAudio[(start + i) % 90000];
     }
